Switch from String to BigInteger. Some housekeeping.

+++ src/main/java/io/oneinch/sdk/model/QuoteResponse.java
@@ -22,5 +22,5 @@ public class QuoteResponse {
     private List<List<List<SelectedProtocol>>> protocols;
     
     @JsonProperty("gas")
-    private Long gas;
+    private BigInteger gas;
 }
